# doc/api.yaml
openapi: 3.0.3
info:
  title: API WASAtext
  description: |-
    API necessarie a WASAText.
  version: "1.0.0"
...
paths:
  /check/{chat}:
    parameters:
    - name: chat
      in: path
      required: true
      description: ID del gruppo
      schema:
        type: integer
        example: 2
        minimum: 1
        maximum: 1024
    get:
      tags: ["Creazione Gruppo"]
      summary: "Controlla se una chat è un gruppo (isGroup)"
      description: |-
        Restituisce true se la chat nell'url è un gruppo altrimenti false
      operationId: isGroup
      responses:
        "200":
          description: Risposta visualizzata correttamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaIsGroup'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: chat non trovata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
...
  /wasachat/{utente}/chats/gruppi/{chat}/aggiungi:
    parameters:
      - name: utente
        in: path
        required: true
        description: Nickname dell'utente
        schema:
          type: string
          example: "Carlo"
          minLength: 1
          maxLength: 50
          pattern: '^[a-zA-Z0-9._~-]+$'


      - name: chat
        in: path
        required: true
        description: ID del gruppo
        schema:
          type: integer
          example: 1
          minimum: 1
          maximum: 1024
    put:
      tags: ["Gestione Gruppi"]
      summary: Aggiungi utente al gruppo (addToGroup)
      description: |-
        Aggiunge un utente al gruppo specificato.
      operationId: addToGroup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RichiestaAggiungiUtente'
        required: true
      responses:
        "200":
          description: Utente aggiunto con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaGenerica'
        "202":
          description: Utente già presente nel gruppo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Utente o gruppo non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
  /wasachat/{utente}/chats/{chat}:
    parameters:
      - name: utente
        in: path
        required: true
        description: Nickname dell'utente
        schema:
          type: string
          example: "Carlo"
          minLength: 1
          maxLength: 50
          pattern: '^[a-zA-Z0-9._~-]+$'


      - name: chat
        in: path
        required: true
        description: ID della chat
        schema:
          type: integer
          example: 1
          minimum: 1
          maximum: 1024
    get:
      tags: ["Chat"]
      summary: Ottieni messaggi chat (getConversation)
      description: |-
        Restituisce i messaggi della chat specificata.
      operationId: getConversation
      responses:
        "200":
          description: Messaggi restituiti con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaMessaggiChat'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Chat non trovata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
    post:
      tags: ["Chat"]
      summary: Invia messaggio (sendMessage)
      description: |-
        Invia un messaggio nella chat specificata.
      operationId: sendMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RichiestaMessaggio'
        required: true
      responses:
        "200":
          description: Messaggio inviato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaGenerica'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Chat non trovata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
    delete:
      tags: ["Gestione Gruppi"]
      summary: Esci dal gruppo (leaveGroup)
      description: |-
        L'utente esce dal gruppo specificato.
      operationId: leaveGroup
      responses:
        "200":
          description: Uscita dal gruppo avvenuta con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaGenerica'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Gruppo non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
  /wasachat/{utente}/inoltro/{nuovachat}/messaggi/{messaggio}:
    parameters:
      - name: utente
        in: path
        required: true
        description: Nickname dell'utente
        schema:
          type: string
          example: "Carlo"
          minLength: 1
          maxLength: 50
          pattern: '^[a-zA-Z0-9._~-]+$'


      - name: nuovachat
        in: path
        required: true
        description: ID della chat di destinazione
        schema:
          type: integer
          example: 2
          minimum: 1
          maximum: 1024
      - name: messaggio
        in: path
        required: true
        description: ID del messaggio da inoltrare
        schema:
          type: integer
          example: 2
          minimum: 1
          maximum: 1024
    post:
      tags: ["Chat"]
      summary: Inoltra messaggio (forwardMessage)
      description: |-
        Inoltra un messaggio a un'altra chat.
      operationId: forwardMessage
      responses:
        "200":
          description: Messaggio inoltrato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaGenerica'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Messaggio o chat non trovati
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
  /inoltro/{utente}/a/{destinatario}/inoltro/messaggi/{messaggio}:
    parameters:
      - name: utente
        in: path
        required: true
        description: Nickname dell'utente inviante
        schema:
          type: string
          example: "Carlo"
          minLength: 1
          maxLength: 50
          pattern: '^[a-zA-Z0-9._~-]+$'
      - name: destinatario
        in: path
        required: true
        description: Nickname dell'utente, ricevente
        schema:
          type: string
          example: "Carlo"
          minLength: 1
          maxLength: 50
          pattern: '^[a-zA-Z0-9._~-]+$'
      - name: messaggio
        in: path
        required: true
        description: ID del messaggio da inoltrare
        schema:
          type: integer
          example: 2
          minimum: 1
          maximum: 1024
    post:
      tags: ["Chat"]
      summary: Inoltra messaggio a una nuova chat (forwardMessageToNewChat)
      description: |-
        Inoltra un messaggio ad una nuova chat, una chat non esistente.
      operationId: forwardMessageToNewChat
      responses:
        "200":
          description: Messaggio inoltrato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaGenerica'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Messaggio o chat non trovati
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
...
  /wasachat/{utente}/gruppi:
    parameters:
    - name: utente
      in: path
      required: true
      description: Nickname dell'utente
      schema:
        type: string
        example: "Carlo"
        minLength: 1
        maxLength: 50
        pattern: '^[a-zA-Z0-9._~-]+$'
    post:
      tags: ["Creazione Gruppo"]
      summary: Crea un gruppo (createGroup)
      description: |-
        Modifica il nome del gruppo specificato.
      operationId: createGroup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RichiestaNomeGruppo'
        required: true
      responses:
        "200":
          description: Nome gruppo modificato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaGenerica'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Gruppo non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
  /wasachat/{utente}/gruppi/{chat}/nome:
    parameters:
    - name: utente
      in: path
      required: true
      description: Nickname dell'utente
      schema:
        type: string
        example: "Carlo"
        minLength: 1
        maxLength: 50
        pattern: '^[a-zA-Z0-9._~-]+$'
    - name: chat
      in: path
      required: true
      description: ID del gruppo
      schema:
        type: integer
        example: 2
        minimum: 1
        maximum: 1024
    put:
      tags: ["Gestione Gruppi"]
      summary: Modifica nome gruppo (setGroupName)
      description: |-
        Modifica il nome del gruppo specificato.
      operationId: setGroupName
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RichiestaPerCreazioneGruppo'
        required: true
      responses:
        "200":
          description: Nome gruppo modificata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaGenerica'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Gruppo non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
  /wasachat/{utente}/gruppi/{chat}/foto:
    parameters:
    - name: utente
      in: path
      required: true
      description: Nickname dell'utente
      schema:
        type: string
        example: "Carlo"
        minLength: 1
        maxLength: 50
        pattern: '^[a-zA-Z0-9._~-]+$'
    - name: chat
      in: path
      required: true
      description: ID del gruppo
      schema:
        type: integer
        example: 2
        minimum: 1
        maximum: 1024
    put:
      tags: ["Gestione Gruppi"]
      summary: Modifica foto gruppo (setGroupPhoto)
      description: |-
        Modifica la foto del gruppo specificato.
      operationId: setGroupPhoto
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RichiestaFotoGruppo'
        required: true
      responses:
        "200":
          description: Foto gruppo modificata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaGenerica'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Gruppo non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
components:
  schemas:
...
    RispostaIsGroup:
      type: object
      description: Indica se la conversazione è di gruppo e opzionalmente l'ID del gruppo se effettivamente è un gruppo
      required:
        - is_group
      properties:
        is_group:
          type: boolean
          description: Indica se la conversazione è un gruppo
        group_id:
          type: integer
          format: int64
          description: ID del gruppo (presente solo se è un gruppo
          minimum: 1
          maximum: 1024
...
    RichiestaAggiungiUtente:
      type: object
      description: Richiesta per aggiungere un utente a un gruppo
      required:
        - utente_da_aggiungere
      properties:
        utente_da_aggiungere:
          type: string
          description: Nickname dell'utente da aggiungere
          example: "Luigi"
          minLength: 1
          maxLength: 1024
          pattern: '^.*$'
    RispostaMessaggiChat:
      type: object
      description: Elenco dei messaggi in una chat
      required:
        - messages
      properties:
        messages:
          type: array
          description: Lista dei messaggi
          minItems: 0
          maxItems: 1024
          items:
            $ref: '#/components/schemas/Messaggio'
...
    Messaggio:
      type: object
      description: Dettagli di un singolo messaggio
      required:
        - message_id
        - autore
        - time
        - ricevuto
        - letto
        - commenti
        - inoltrato
      properties:
        message_id:
          type: integer
          description: ID univoco del messaggio
          example: 12
          minimum: 1
          maximum: 1024
        autore:
          type: string
          description: Nickname dell'autore
          example: "Marcolino"
          minLength: 1
          maxLength: 50
          pattern: '^.*$'
        text:
          type: string
          description: Contenuto testuale del messaggio
          example: "Sei libero sabato?"
          minLength: 1
          maxLength: 1000
          pattern: '^.*$'
        foto:
          type: string
          description: Contenuto base64 della foto
          example: "fasfgasdkjfvashfadsfvhjfasfbashbvdfj--"
          minLength: 1
          maxLength: 1024
          pattern: '^.*$'
        time:
          type: string
          format: date-time
          description: Timestamp del messaggio in formato ISO 8601
          example: "2025-03-15T14:30:00Z"
          minLength: 20
          maxLength: 30
          pattern: '^.*$'
        ricevuto:
          type: boolean
          description: Indica se il messaggio è stato ricevuto
          example: true
        letto:
          type: boolean
          description: Indica se il messaggio è stato letto
          example: false
        commenti:
          type: array
          description: Lista dei commenti associati al messaggio
          minItems: 0
          maxItems: 100
          items:
            $ref: '#/components/schemas/Commento'
        inoltrato:
          type: boolean
          description: Indica se il messaggio è stato inoltrato
          example: true
        risposta:
          type: integer
          description: ID del messaggio a cui si sta rispondendo
          example: 9
          minimum: 1
          maximum: 1024
    RichiestaMessaggio:
      type: object
      description: Richiesta per inviare un messaggio
      properties:
        testo:
          type: string
          description: Contenuto testuale del messaggio
          example: "Ti va di uscire insieme questo venerdì?"
          minLength: 1
          maxLength: 1000
          pattern: '^.*$'
        foto:
          type: string
          description: Contenuto base64 della foto
          example: "fasfgasdkjfvashfadsfvhjfasfbashbvdfj--"
          minLength: 1
          maxLength: 1024
          pattern: '^.*$'
...
    Commento:
      type: object
      description: Dettagli di un singolo commento
      required:
        - comment_id
        - autore
        - reazione
      properties:
        comment_id:
          type: integer
          description: ID univoco del commento
          example: 2
          minimum: 1
          maximum: 1024
        autore:
          type: string
          description: Nickname dell'autore
          example: "Marcolino"
          minLength: 1
          maxLength: 50
          pattern: '^.*$'
        reazione:
          type: string
          description: Contenuto del commento (emoji o testo breve)
          example: "👍"
          minLength: 1
          maxLength: 10
          pattern: '^[\u0020-\u1FAFF]+$'
...
    RichiestaPerCreazioneGruppo:
      type: object
      description: Richiesta per creare gruppi
      required:
        - nome
        - foto
      properties:
        nome:
          type: string
          description: nome
          example: "Peppe idraulico"
          minLength: 1
          maxLength: 100
          pattern: '^.*$'
        foto:
          type: string
          description: Contenuto base64 della foto
          example: "fasfgasdkjfvashfadsfvhjfasfbashbvdfj--"
          minLength: 1
          maxLength: 1024
          pattern: '^.*$'
    RichiestaNomeGruppo:
      type: object
      description: Richiesta per modificare il nome di un gruppo
      required:
        - nome
      properties:
        nome:
          type: string
          description: nome
          example: "Peppe idraulico"
          minLength: 1
          maxLength: 100
          pattern: '^.*$'
    RichiestaFotoGruppo:
      type: object
      description: Richiesta per modificare la foto di un gruppo
      required:
        - foto
      properties:
        foto:
          type: string
          description: Contenuto base64 della foto
          example: "fasfgasdkjfvashfadsfvhjfasfbashbvdfj--"
          minLength: 1
          maxLength: 1024
          pattern: '^.*$'
    RispostaGenerica:
      type: object
      description: Risposta generica
      required:
        - codice
        - risposta
      properties:
        codice:
          type: integer
          description: codice di risposta
          example: 200
          minLength: 1
          maxLength: 1024
          minimum: 200
          maximum: 299
        risposta:
          type: string
          description: risposta 
          example: "messaggio di successo"
          minLength: 1
          maxLength: 1024
          pattern: '^.*$'
    ErroreGenerico:
      type: object
      description: Errore
      required:
        - codice
        - errore
      properties:
        codiceErrore:
          type: integer
          description: codice di errore
          example: 500
          minLength: 1
          maxLength: 1024
          minimum: 300
          maximum: 599
        errore:
          type: string
          description: errore
          pattern: '^.*$'
          example: "errore"
          minLength: 1
          maxLength: 1024